prepare channel sorting in mark.initialize
