+++ src/scales/ordinal.js
@@ -1,6 +1,6 @@
 import {InternSet, rollup, reverse as reverseof, sort} from "d3";
-import {range} from "../mark.js";
 import {scaleBand, scaleOrdinal, scalePoint} from "d3";
+import {range} from "../mark.js";
 import {ordinalScheme} from "./schemes.js";
 import {ascendingDefined} from "../defined.js";
 import {registry, color} from "./index.js";
@@ -75,9 +75,7 @@ function inferDomain(channels) {
   for (const {value, sorted} of channels) {
     if (value === undefined) continue;
     for (const v of value) domain.add(v);
-    if (sorted) {
-      a = rollup(range(value), sorted, i => value[i]);
-    }
+    if (sorted) a = rollup(range(value), sorted, i => value[i]); // TODO rename a
   }
   return sort(domain, a ? v => a.get(v) : ascendingDefined);
 }
